+++ lib/nxt_support.rb
@@ -7,6 +7,10 @@ require "nxt_support/models"
 require "nxt_support/serializers"
 require "nxt_support/util"
 require "nxt_support/services"
+require "nxt_support/refinements"
+
+require 'nxt_support/middleware/sentry_error_id'
+
 
 module NxtSupport
 end